+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 이 프로젝트의 변경 사항을 문서화합니다.
 
+## [1.0.5] - 2024-12-23
+
+### Added
+
+- [maskName](src/masking/maskName.ts) 함수 추가
+- [maskPhoneNumber](src/masking/maskPhoneNumber.ts) 함수 추가
+- [commaizeNumber](src/number/commaizeNumber.ts) 함수 추가
+- [decommaizeNumber](src/number/decommaizeNumber.ts) 함수 추가
+- [floorAndFormatNumber](src/number/floorAndFormatNumber.ts) 함수 추가
+- [ceilToUnit](src/number/formatter.ts) 함수 추가
+- [floorToUnit](src/number/formatter.ts) 함수 추가
+- [roundToUnit](src/number/formatter.ts) 함수 추가
+- [formatToKoreanNumber](src/number/formatToKoreanNumber.ts) 함수 추가
+- [formatToKRW](src/number/formatToKRW.ts) 함수 추가
+- [decode](src/decode.ts) 함수 추가
+
 ## [1.0.4] - 2024-12-09
 
 ### Added